requirements.txt
streamlit
langchain
langchain-core
langchain-community
langchain-huggingface
huggingface-hub
python-dotenv
pypdf